// Soe.Angular.Spa/src/app/features/manage/sys-edi-message-head/models/sys-edi-message-head.model.ts
export class SysEdiMessageHeadDTO implements ISysEdiMessageHeadDTO {
  sysEdiMessageHeadId!: number;
  sysEdiMessageHeadGuid!: string;
  sysEdiMessageHeadStatus!: SysEdiMessageHeadStatus;
  sysCompanyId?: number;
  sysEdiMessageRawId!: number;
  ediSourceType!: number;
  sysEdiType?: number;
  sysEdiMsgId?: number;
  sysWholesellerId!: number;
  messageSenderId!: string;
  messageType!: string;
  messageDate?: Date;
  sellerId!: string;
  sellerOrganisationNumber!: string;
  sellerVatNumber!: string;
  sellerName!: string;
  sellerAddress!: string;
  sellerPostalCode!: string;
  sellerPostalAddress!: string;
  sellerCountryCode!: string;
  sellerPhone!: string;
  sellerFax!: string;
  sellerReference!: string;
  sellerReferencePhone!: string;
  buyerId!: string;
  buyerOrganisationNumber!: string;
  buyerVatNumber!: string;
  buyerName!: string;
  buyerAddress!: string;
  buyerPostalCode!: string;
  buyerPostalAddress!: string;
  buyerCountryCode!: string;
  buyerReference!: string;
  buyerPhone!: string;
  buyerFax!: string;
  buyerEmailAddress!: string;
  buyerDeliveryName!: string;
  buyerDeliveryCoAddress!: string;
  buyerDeliveryAddress!: string;
  buyerDeliveryPostalCode!: string;
  buyerDeliveryPostalAddress!: string;
  buyerDeliveryCountryCode!: string;
  buyerDeliveryNoteText!: string;
  buyerDeliveryGoodsMarking!: string;
  headInvoiceNumber!: string;
  headInvoiceOcr!: string;
  headInvoiceType!: string;
  headInvoiceDate?: Date;
  headInvoiceDueDate?: Date;
  headDeliveryDate?: Date;
  headBuyerOrderNumber!: string;
  headBuyerProjectNumber!: string;
  headBuyerInstallationNumber!: string;
  headSellerOrderNumber!: string;
  headPostalGiro!: string;
  headBankGiro!: string;
  headBank!: string;
  headBicAddress!: string;
  headIbanNumber!: string;
  headCurrencyCode!: string;
  headVatPercentage?: number;
  headPaymentConditionDays?: number;
  headPaymentConditionText!: string;
  headInterestPaymentPercent!: string;
  headInterestPaymentText!: string;
  headInvoiceGrossAmount!: number;
  headInvoiceNetAmount!: number;
  headVatBasisAmount!: number;
  headVatAmount!: number;
  headFreightFeeAmount!: number;
  headHandlingChargeFeeAmount!: number;
  headInsuranceFeeAmount!: number;
  headRemainingFeeAmount!: number;
  headDiscountAmount!: number;
  headRoundingAmount!: number;
  headBonusAmount!: number;
  headInvoiceArrival!: string;
  headInvoiceAuthorized!: string;
  headInvoiceAuthorizedBy!: string;
  xDocument!: string;
  errorMessage!: string;
  created?: Date;
  lastSendTry?: Date;
  sent?: Date;
  sysEdiEdiMessageRowDTOs!: ISysEdiMessageRowDTO[];
}
// ...
export enum SysEdiMessageHeadStatus {
  Unknown = 0,
  Unhandled = 1,
  Handled = 2,
  NoSysCompanyFound = 5,
  SentToComp = 10,
  Error = 99,
}

export interface ISysEdiMessageHeadDTO {
  sysEdiMessageHeadId: number;
  sysEdiMessageHeadGuid: string;
  sysEdiMessageHeadStatus: SysEdiMessageHeadStatus;
  sysCompanyId?: number;
  sysEdiMessageRawId: number;
  ediSourceType: number;
  sysEdiType?: number;
  sysEdiMsgId?: number;
  sysWholesellerId: number;
  messageSenderId: string;
  messageType: string;
  messageDate?: Date;
  sellerId: string;
  sellerOrganisationNumber: string;
  sellerVatNumber: string;
  sellerName: string;
  sellerAddress: string;
  sellerPostalCode: string;
  sellerPostalAddress: string;
  sellerCountryCode: string;
  sellerPhone: string;
  sellerFax: string;
  sellerReference: string;
  sellerReferencePhone: string;
  buyerId: string;
  buyerOrganisationNumber: string;
  buyerVatNumber: string;
  buyerName: string;
  buyerAddress: string;
  buyerPostalCode: string;
  buyerPostalAddress: string;
  buyerCountryCode: string;
  buyerReference: string;
  buyerPhone: string;
  buyerFax: string;
  buyerEmailAddress: string;
  buyerDeliveryName: string;
  buyerDeliveryCoAddress: string;
  buyerDeliveryAddress: string;
  buyerDeliveryPostalCode: string;
  buyerDeliveryPostalAddress: string;
  buyerDeliveryCountryCode: string;
  buyerDeliveryNoteText: string;
  buyerDeliveryGoodsMarking: string;
  headInvoiceNumber: string;
  headInvoiceOcr: string;
  headInvoiceType: string;
  headInvoiceDate?: Date;
  headInvoiceDueDate?: Date;
  headDeliveryDate?: Date;
  headBuyerOrderNumber: string;
  headBuyerProjectNumber: string;
  headBuyerInstallationNumber: string;
  headSellerOrderNumber: string;
  headPostalGiro: string;
  headBankGiro: string;
  headBank: string;
  headBicAddress: string;
  headIbanNumber: string;
  headCurrencyCode: string;
  headVatPercentage?: number;
  headPaymentConditionDays?: number;
  headPaymentConditionText: string;
  headInterestPaymentPercent: string;
  headInterestPaymentText: string;
  headInvoiceGrossAmount: number;
  headInvoiceNetAmount: number;
  headVatBasisAmount: number;
  headVatAmount: number;
  headFreightFeeAmount: number;
  headHandlingChargeFeeAmount: number;
  headInsuranceFeeAmount: number;
  headRemainingFeeAmount: number;
  headDiscountAmount: number;
  headRoundingAmount: number;
  headBonusAmount: number;
  headInvoiceArrival: string;
  headInvoiceAuthorized: string;
  headInvoiceAuthorizedBy: string;
  xDocument: string;
  errorMessage: string;
  created?: Date;
  lastSendTry?: Date;
  sent?: Date;
  sysEdiEdiMessageRowDTOs: ISysEdiMessageRowDTO[];
}

export interface ISysEdiMessageRowDTO {
  sysEdiMessageRowId: number;
  sysEdiMessageHeadId: number;
  rowSellerArticleNumber: string;
  rowSellerArticleDescription1: string;
  rowSellerArticleDescription2: string;
  rowSellerRowNumber: string;
  rowBuyerArticleNumber: string;
  rowBuyerRowNumber: string;
  rowDeliveryDate?: Date;
  rowBuyerReference: string;
  rowBuyerObjectId: string;
  rowQuantity?: number;
  rowUnitCode: string;
  rowUnitPrice: string;
  rowDiscountPercent?: number;
  rowDiscountAmount?: number;
  rowDiscountPercent1?: number;
  rowDiscountAmount1?: number;
  rowDiscountPercent2?: number;
  rowDiscountAmount2?: number;
  rowNetAmount?: number;
  rowVatAmount?: number;
  rowVatPercentage?: number;
  externalProductId: number;
  stockCode: string;
}